Architecture: moved Connect4Board and TicTacToe java
